Add files via upload
I recently acquired an analog clock that is quite unique. It not only displays the time in analog format, but it also has two boxes within it. The first box displays the time in digital format, while the second box shows the day of the week. It's like having a brand new analog clock!